move RevisionsView to a separate module
